silence all -Wsign-compare warnings (use size_t instead of int in a few places)

+++ src/Symbol/Variable.h
@@ -1,12 +1,14 @@
 #ifndef _XMUTIL_SYMBOL_VARIABLE_H
 #define _XMUTIL_SYMBOL_VARIABLE_H
+#include <cstddef>
+#include <vector>
+
 #include "Symbol.h"
 #include "Equation.h"
 #include "UnitExpression.h"
 #include "../Function/State.h"
 #include "../Function/Function.h"
 #include "../XMUtil.h"
-#include <vector>
 class View;
 
 /* varibale content - the true stuff for variables without the associated symbol
@@ -147,8 +149,8 @@ public:
    void MarkAsFlow() { bAsFlow = true; }
    bool AsFlow() const { return bAsFlow; }
 
-   int Nelm() const { return iNelm; }
-   void SetNelm(int set) { iNelm = set; }
+   size_t Nelm() const { return iNelm; }
+   void SetNelm(size_t set) { iNelm = set; }
 
    // for flowing
    bool HasUpstream() const { return _hasUpstream; }
@@ -168,7 +170,7 @@ private :
 	std::vector<Variable*> mOutflows;
    VariableContent *pVariableContent ; // dependent on variable type which is not known on instantiation
    XMILE_Type mVariableType;
-   int iNelm; // used for subscript owners
+   size_t iNelm; // used for subscript owners
    View* _view; // view defined in
    bool _unwanted;
    bool _hasUpstream;
